+++ app/models/unidom/shopping/concerns/as_cart_shopper.rb
@@ -6,6 +6,14 @@ module Unidom::Shopping::Concerns::AsCartShopper
 
     has_many :shopping_carts, class_name: 'Unidom::Shopping::ShoppingCart', as: :shopper
 
+    def get_cart!(from: nil, at: Time.now)
+      shopping_carts.shop_is(from).valid_at(now: at).alive.first_or_create! opened_at: at
+    end
+
+    #def get_cart?(from: nil, at: Time.now)
+    #  shopping_carts.shop_is(from).valid_at(now: at).alive.exists?
+    #end
+
   end
 
   module ClassMethods